+++ app/src/main/java/com/skyyo/samples/features/exoPlayer/VideoPlayerWithControls.kt
@@ -4,7 +4,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.widget.ImageButton
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
@@ -13,13 +12,12 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.viewinterop.AndroidView
-import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.SimpleExoPlayer
 import com.google.android.exoplayer2.ui.PlayerView
 import com.skyyo.samples.R
 
 @Composable
-fun VideoPlayerWithControls(exoPlayer: ExoPlayer) {
+fun VideoPlayerWithControls(exoPlayer: SimpleExoPlayer) {
     val context = LocalContext.current
     val playerView = remember {
         val layout = LayoutInflater.from(context).inflate(R.layout.video_player_auto, null)
@@ -32,5 +30,5 @@ fun VideoPlayerWithControls(exoPlayer: ExoPlayer) {
         playerView
     }
 
-    AndroidView({ playerView }, Modifier.fillMaxSize().background(Color.Black))
+    AndroidView({ playerView }, Modifier.height(256.dp).background(Color.Black))
 }
