+++ app/src/main/java/com/skyyo/samples/features/verticalPagerWithFling/composables/FullscreenVideoPlayer.kt
@@ -1,0 +1,26 @@
+package com.skyyo.samples.features.verticalPagerWithFling.composables
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.viewinterop.AndroidView
+import com.google.android.exoplayer2.ExoPlayer
+import com.google.android.exoplayer2.ui.AspectRatioFrameLayout
+import com.google.android.exoplayer2.ui.StyledPlayerView
+
+@Composable
+fun FullscreenVideoPlayer(exoPlayer: ExoPlayer) {
+    val context = LocalContext.current
+
+    AndroidView(
+        factory = {
+            StyledPlayerView(context).apply {
+                useController = false
+                resizeMode = AspectRatioFrameLayout.RESIZE_MODE_ZOOM
+                player = exoPlayer
+            }
+        },
+        modifier = Modifier.fillMaxSize()
+    )
+}
